% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-snippets.R
\name{bootstrap_snippet_loader}
\alias{bootstrap_snippet_loader}
\title{Return a function that loads snippet files from the given directories.
This will be used to create a \code{snippets()} function in the episode build environment.}
\usage{
bootstrap_snippet_loader(custom_snippets, base_snippets = NULL)
}
\arguments{
\item{custom_snippets}{the path to the custom snippets directory (or \code{NULL} if none)}

\item{base_snippets}{the path to the base snippets directory (or \code{NULL} if none)}
}
\value{
a function that takes a child file path and renders it as a snippet, or returns the path to the snippet file if \code{render = FALSE}.
}
\description{
Return a function that loads snippet files from the given directories.
This will be used to create a \code{snippets()} function in the episode build environment.
}
\keyword{internal}
