% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_404.R
\name{build_404}
\alias{build_404}
\title{Build the 404 page for a lesson}
\usage{
build_404(pkg, quiet = FALSE)
}
\arguments{
\item{pkg}{a list object generated from \code{\link[pkgdown:as_pkgdown]{pkgdown::as_pkgdown()}}}

\item{quiet}{passed to \code{\link[=build_html]{build_html()}}. When \code{FALSE} (default), a message
will be printed to the screen about the build progress. When \code{TRUE}, no
messages are generated.}
}
\value{
\code{TRUE} if the page was successfully generated
}
\description{
Build the 404 page for a lesson
}
\details{
During the lesson build process, a 404 page with absolute links back to the
source pages must be generated otherwise, subsequent attempts to escape the
404 page will be futile.

This function is intended to be run on a lesson website that has already
been built and is called for its side-effect of creating a 404 page.
}
\seealso{
\code{\link[=build_site]{build_site()}} which calls this function and \code{\link[=build_html]{build_html()}}, which
this function calls.
}
\keyword{internal}
