% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_aio.R
\name{build_aio}
\alias{build_aio}
\title{Build the All-in-one page}
\usage{
build_aio(pkg, quiet)
}
\arguments{
\item{pkg}{an object created by {pkgdown}, supplied by \code{\link[=build_site]{build_site()}}.}

\item{quiet}{If \code{TRUE} then no messages will be shown when building.

This function will build the all-in-one page for the lesson website. Because
the bottleneck is often the internal processes of {pkgdown}, we are first
templating the AIO page as a blank page and then adding in the contents using
{xml2}.}
}
\description{
Build the All-in-one page
}
\note{
This function assumes that you have already built all the episodes of your
lesson.
}
\examples{
if (FALSE) {
  # build_aio() assumes that your lesson has been built and takes in a 
  # pkgdown object, which can be created from the `site/` folder in your
  # lesson.
  lsn <- "/path/to/my/lesson"
  pkg <- pkgdown::as_pkgdown(fs::path(lsn, "site"))

  build_aio(pkg, quiet = FALSE)
}
}
\seealso{
\link{provision_aio}, \link{make_section}, \link{get_sections}
}
\keyword{internal}
