% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_citation.R
\name{build_citation}
\alias{build_citation}
\title{Create a citation page for a lesson from a CITATION.CFF file}
\usage{
build_citation(pkg, quiet = FALSE)
}
\arguments{
\item{pkg}{an object created via \code{\link[pkgdown:as_pkgdown]{pkgdown::as_pkgdown()}} of a lesson.}

\item{quiet}{if \code{TRUE}, no messages will be emitted. If FALSE, pkgdown will
report creation of the temporary file.}
}
\value{
NULL, invisibly. This is called for its side-effect
}
\description{
Create a citation page for a lesson from a CITATION.CFF file
}
\keyword{internal}
