% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_episode.R
\name{build_episode_html}
\alias{build_episode_html}
\title{Build a single episode html file}
\usage{
build_episode_html(
  path_md,
  path_src = path_md,
  page_back = NULL,
  page_forward = NULL,
  pkg,
  quiet = FALSE
)
}
\arguments{
\item{path_md}{the path to the episode markdown (not RMarkdown) file
(usually via \code{\link[=build_episode_md]{build_episode_md()}}).}

\item{path_src}{the path to the source file (defaults \code{path_md})}

\item{page_back}{the URL for the previous page}

\item{page_forward}{the URL for the next page}

\item{pkg}{a \code{pkgdown} object containing metadata for the site}

\item{quiet}{if \code{TRUE}, messages are not produced. Defaults to \code{TRUE}.}
}
\value{
\code{TRUE} if the page was successful, \code{FALSE} otherwise.
}
\description{
Build a single episode html file
}
\note{
this function is for internal use, but exported for those who know what
they are doing.
}
\examples{
tmp <- tempfile()
create_lesson(tmp, open = FALSE)
suppressWarnings(set_schedule(tmp, get_schedule(tmp), write = TRUE))
build_lesson(tmp)

# create a new file in extras
fun_file <- file.path(tmp, "episodes", "extras", "fun.Rmd")
txt <- c(
 "---\ntitle: Fun times\n---\n\n",
 "# new page\n", 
 "This is coming from `r R.version.string`"
)
file.create(fun_file)
writeLines(txt, fun_file)
hash <- tools::md5sum(fun_file)
res <- build_episode_md(fun_file, hash)
build_episode_html(res, 
  page_back = "index.md",
  page_forward = "index.md",
  pkg = pkgdown::as_pkgdown(file.path(tmp, "site"))
)
}
\seealso{
\code{\link[=build_episode_md]{build_episode_md()}}, \code{\link[=build_lesson]{build_lesson()}}, \code{\link[=build_markdown]{build_markdown()}}
}
\keyword{internal}
