% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_episode.R
\name{build_episode_ipynb}
\alias{build_episode_ipynb}
\title{Convert a single episode to a Jupyter notebook}
\usage{
build_episode_ipynb(
  path,
  outdir = path_built(path),
  profile = "lesson-requirements",
  quiet = TRUE
)
}
\arguments{
\item{path}{path to the RMarkdown file.}

\item{outdir}{the directory to write to/}

\item{quiet}{if \code{TRUE}, output is suppressed, default is \code{TRUE} so we can use better
formatting elsewhere.}
}
\value{
the path to the output, invisibly
}
\description{
Convert a single episode to a Jupyter notebook
}
\keyword{internal}
