% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_episode.R
\name{build_episode_md}
\alias{build_episode_md}
\title{Build an episode to markdown}
\usage{
build_episode_md(
  path,
  hash,
  outdir = path_built(path),
  workdir = path_episodes(path),
  env = new.env(),
  quiet = FALSE
)
}
\arguments{
\item{path}{path to the RMarkdown file}

\item{hash}{hash to prepend to the output}

\item{outdir}{the directory to write to}

\item{workdir}{the directory where the episode should be rendered}

\item{env}{a blank environment}

\item{quiet}{if \code{TRUE}, output is suppressed, default is \code{FALSE} to show
{knitr} output.}
}
\value{
the path to the output, invisibly
}
\description{
This uses \code{\link[knitr:knit]{knitr::knit()}} with custom options set for the Carpentries
template and prepends a hash to the yaml header
}
\note{
this function is for internal use, but exported for those who know what
they are doing.
}
\examples{
fun_dir <- tempfile()
dir.create(fun_dir)
fun_file <- file.path(fun_dir, "fun.Rmd")
file.create(fun_file)
txt <- c(
 "---\ntitle: Fun times\n---\n\n",
 "# new page\n", 
 "This is coming from `r R.version.string`"
)
writeLines(txt, fun_file)
hash <- tools::md5sum(fun_file)
res <- build_episode_md(fun_file, hash, outdir = fun_dir, workdir = fun_dir)
}
\seealso{
\code{\link[=build_episode_html]{build_episode_html()}}
}
\keyword{internal}
