% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_glossary.R
\name{build_glossary}
\alias{build_glossary}
\alias{build_glossary_page}
\title{Build a page for aggregating glosario links found in lesson elements}
\usage{
build_glossary(pkg, pages = NULL, quiet = FALSE)

build_glossary_page(
  pkg,
  pages,
  title = "Glosario Links",
  slug = "reference",
  aggregate = "*",
  append = "self::node()",
  quiet = FALSE
)
}
\arguments{
\item{pkg}{an object created via \code{\link[pkgdown:as_pkgdown]{pkgdown::as_pkgdown()}} of a lesson.}

\item{pages}{output from the function \code{\link[=read_all_html]{read_all_html()}}: a nested list of
\code{xml_document} objects representing episodes in the lesson}

\item{quiet}{if \code{TRUE}, no messages will be emitted. If FALSE, pkgdown will
report creation of the temporary file.}

\item{title}{the new page title}

\item{slug}{the slug for the page (e.g. "aio" will become "aio.html")}

\item{aggregate}{a selector for the lesson content you want to aggregate.
The default is "*", which will aggregate links from all content.
To grab only links from sections, use "section".}

\item{append}{a selector for the section of the page where the aggregate data
should be placed. This defaults to "self::node()", which indicates that the
entire page should be appended.}
}
\value{
NULL, invisibly. This is called for its side-effect
}
\description{
Build a page for aggregating glosario links found in lesson elements
}
\details{
We programmatically search through lesson content to find links that point to
glosario terms. We then aggregate these links into the Reference.
}
\note{
This function assumes that you have already built all the episodes of your lesson.
}
\examples{
if (FALSE) {
  # build_glossary_page() assumes that your lesson has been built and takes in a
  # pkgdown object, which can be created from the `site/` folder in your
  # lesson.
  lsn <- "/path/to/my/lesson"
  pkg <- pkgdown::as_pkgdown(fs::path(lsn, "site"))

  htmls <- read_all_html(pkg$dst_path)
  build_glossary_page(pkg, htmls, quiet = FALSE)
}
}
\keyword{internal}
