% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_handout.R
\name{build_handout}
\alias{build_handout}
\title{Create a code handout of challenges without solutions}
\usage{
build_handout(path = ".")
}
\arguments{
\item{path}{the path to the lesson. Defaults to current working directory}
}
\description{
Create a code handout of challenges without solutions
}
