% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_home.R
\name{build_home}
\alias{build_home}
\title{Build a home page for a lesson}
\usage{
build_home(pkg, quiet, next_page = NULL)
}
\arguments{
\item{pkg}{a list generated from \code{\link[pkgdown:as_pkgdown]{pkgdown::as_pkgdown()}} from the \verb{site/}
folder of a lesson.}

\item{quiet}{a boolean passed to \code{\link[=build_html]{build_html()}}. if \code{TRUE}, this will have
pkgdown report what files are being built}

\item{next_page}{the next page file name. This will allow the navigation
element to be set up correctly on the navigation bar}
}
\value{
nothing. This is used for its side-effect
}
\description{
Build a home page for a lesson
}
\details{
The index page of the lesson is a combination of two pages:
\enumerate{
\item index.md (or README if the index does not exist)
\item learners/setup.md
}

This function uses \code{\link[=render_html]{render_html()}} to convert the page into HTML, which gets
passed on to the "syllabus" or "overview" templates in \code{{varnish}} (via the
\code{\link[=build_html]{build_html()}} function as the \code{{{{ readme }}}} and \code{{{{ setup }}}} keys.
}
\keyword{internal}
