% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_html.R
\name{build_html}
\alias{build_html}
\title{Build instructor and learner HTML page}
\usage{
build_html(
  template = "chapter",
  pkg,
  nodes,
  global_data,
  path_md,
  quiet = TRUE
)
}
\arguments{
\item{template}{the name of the {varnish} template to use. Defaults to
"chapter"}

\item{pkg}{an object created from  \code{\link[pkgdown:as_pkgdown]{pkgdown::as_pkgdown()}}}

\item{nodes}{an \code{xml_document} object. In the case of using this for the
index page (from \code{\link[=build_home]{build_home()}}), \code{nodes} will be a list of two
\code{xml_documents}; one for instructors and one for learners so that the
instructors have the schedule available to them (however, this might be a
relic, Zhian needs to inspect it).}

\item{global_data}{a list store object that contains copies of the global
variables for the page, including metadata, navigation, and variables for
the {varnish} templates.}

\item{path_md}{the path (absolute, relative, or filename) the current
markdown file being processed.}

\item{quiet}{This parameter is passed to \code{\link[pkgdown:render_page]{pkgdown::render_page()}} and will
print the progress if \code{TRUE} (default).}
}
\value{
\code{TRUE} if the page was built and \code{NULL} if it did not need to be
rebuilt
}
\description{
Build instructor and learner HTML page
}
\details{
This function is a central workhorse that connects the global
lesson metadata and the global variables for each page to the rendering
engine: {pkgdown}. It will perform the global operations that includes
setting up the navigation, adding metadata, and building both the instructor
and learner versions of the page.

In the Workbench, there are three types of pages:
\enumerate{
\item primary content pages: these are primary content with a 1:1 relationship
between the source and the output. These are episodes along with custom
learner and instructor content
\item aggregate content pages: pages that are aggregated from other pages such
as key points, all-in-one, images
\item concatenated content pages: concatenations of source files and potentially
aggregate data. Examples are index, learner profiles, and the instructor
notes pages.
}

Each of these types of pages have their own process for setting up content,
which gets processed before its passed here.
}
\keyword{internal}
