% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ipynb.R
\name{build_ipynb}
\alias{build_ipynb}
\title{Build ipynb notebooks from the RMarkdown episodes}
\usage{
build_ipynb(path = ".", rebuild = FALSE, quiet = FALSE)
}
\arguments{
\item{path}{the path to your repository (defaults to your current working
directory)}

\item{rebuild}{if \code{TRUE}, everything will be built from scratch as if there
was no cache. Defaults to \code{FALSE}, which will only build ipynb files that
haven't been built before.}

\item{quiet}{when \code{TRUE}, output is supressed.}
}
\description{
Convert the RMarkdown episodes to Jupyter notebooks using \code{jupytext}.
Only the actual episodes are converted, not the additional pages.
}
\seealso{
\code{\link[=build_episode_ipynb]{build_episode_ipynb()}}
}
\keyword{internal}
