% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{build_site}
\alias{build_site}
\title{Wrapper for site builder}
\usage{
build_site(
  path = ".",
  quiet = !interactive(),
  preview = TRUE,
  override = list()
)
}
\arguments{
\item{path}{the path to your repository (defaults to your current working
directory)}

\item{quiet}{when \code{TRUE}, output is supressed}

\item{preview}{if \code{TRUE}, the rendered website is opened in a new window}

\item{override}{options to override (e.g. building to alternative paths).
This is used internally.}
}
\description{
Wrapper for site builder
}
\note{
this assumes that the markdown files have already been built and will
not work otherwise
}
\keyword{internal}
