% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pandoc.R
\name{check_pandoc}
\alias{check_pandoc}
\title{Check the existence of pandoc}
\usage{
check_pandoc(quiet = TRUE, pv = "2.11", rv = "1.4")
}
\arguments{
\item{quiet}{if \code{TRUE}, no message will be emitted, otherwise the pandoc
version and path will be sent as a message (stderr) to the screen.}

\item{pv}{the minimum pandoc version}

\item{rv}{the minimum rstudio version (if available)}
}
\description{
This function adds context to \code{\link[rmarkdown:pandoc_available]{rmarkdown::pandoc_available()}} and provides
an error message directing the user to download the latest version of pandoc
or RStudio Desktop.
}
\examples{
# NOTE: this is an internal function, so there is no guarantee that the usage
# will remain the same across time. This is merely for demonstration purposes
# only.

# Check for pandoc ----------------------
asNamespace("sandpaper")$check_pandoc(quiet = FALSE)

# Message emitted when pandoc cannot be found --------
try(asNamespace("sandpaper")$check_pandoc(quiet = FALSE, pv = "999"))
}
\keyword{internal}
