% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_build.R
\name{ci_build_markdown}
\alias{ci_build_markdown}
\alias{ci_build_site}
\title{Build and deploy individual site components to a remote branch}
\usage{
ci_build_markdown(
  path = ".",
  branch = "md-outputs",
  remote = "origin",
  reset = FALSE
)

ci_build_site(
  path = ".",
  branch = "gh-pages",
  md = "md-outputs",
  remote = "origin",
  reset = FALSE
)
}
\arguments{
\item{path}{path to the lesson}

\item{branch}{the branch name containing the output:
\itemize{
\item \code{ci_build_markdown()}: defaults to \code{md-output}, corresponds to
\code{md_branch} in \code{\link[=ci_deploy]{ci_deploy()}}
\item \code{ci_build_site()}: defaults to \code{gh-pages}, corresponds to \code{site_branch}
in \code{\link[=ci_deploy]{ci_deploy()}}
}}

\item{remote}{the name of the git remote to which you should deploy.}

\item{reset}{if \code{TRUE}, the markdown cache is cleared before rebuilding,
this defaults to \code{FALSE} meaning the markdown cache will be provisioned
and used.}

\item{md}{the branch name that contains the markdown outputs}
}
\value{
\itemize{
\item \code{ci_build_markdown()}: an expression that is evaluated to tear down the
worktree that was created.
\item \code{ci_build_site()}: nothing, it is used for its side-effect.
}
}
\description{
These are the internal functions which provision and deploy lesson
components to orphan git branches. These are called by \code{\link[=ci_deploy]{ci_deploy()}}.
\strong{These functions have the side-effect of creating and updating orphan
branches on a remote repository. It is assumed that you have access to your
remote}.
}
\details{
TBW
}
\note{
\code{ci_build_markdown()} will set the \code{sandpaper.use_renv}
option to TRUE, which means that it will \emph{always} use the {renv}
package cache if the lesson uses R Markdown.
}
\keyword{internal}
