% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_build_markdown.R
\name{ci_build_markdown}
\alias{ci_build_markdown}
\title{(INTERNAL) Build and deploy the site with continous integration}
\usage{
ci_build_markdown(path = ".", branch = "md-outputs", remote = "origin")
}
\arguments{
\item{path}{path to the lesson}

\item{remote}{the name of the git remote to which you should deploy.}

\item{md_branch}{the branch name that contains the markdown outputs}

\item{site_branch}{the branch name that contains the full HTML site}
}
\value{
NOTHING
}
\description{
(INTERNAL) Build and deploy the site with continous integration
}
\note{
this function is not for interactive use. It requires git to be
installed on your machine and will destroy anything you have in the
\verb{site/} folder.
}
\keyword{internal}
