% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_build_site.R
\name{ci_build_site}
\alias{ci_build_site}
\title{(INTERNAL) Build and deploy the site with continous integration}
\usage{
ci_build_site(
  path = ".",
  branch = "gh-pages",
  md = "md-outputs",
  remote = "origin"
)
}
\arguments{
\item{path}{path to the lesson}

\item{branch}{the branch name that contains the full HTML site}

\item{md}{the branch name that contains the markdown outputs}

\item{remote}{the name of the git remote to which you should deploy.}
}
\value{
NOTHING
}
\description{
(INTERNAL) Build and deploy the site with continous integration
}
\note{
this function is not for interactive use. It requires git to be
installed on your machine and will destroy anything you have in the
\verb{site/} folder. Additionally, this will set the \code{sandpaper.use_renv}
option to TRUE, which means that it will \emph{always} use the {renv}
package cache.
}
\keyword{internal}
