% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_deploy.R
\name{ci_deploy}
\alias{ci_deploy}
\title{(INTERNAL) Build and deploy the site with continous integration}
\usage{
ci_deploy(
  path = ".",
  md_branch = "md-outputs",
  site_branch = "gh-pages",
  remote = "origin",
  reset = FALSE
)
}
\arguments{
\item{path}{path to the lesson}

\item{md_branch}{the branch name that contains the markdown outputs}

\item{site_branch}{the branch name that contains the full HTML site}

\item{remote}{the name of the git remote to which you should deploy.}

\item{reset}{if \code{TRUE}, the site cache is cleared before rebuilding, this
defaults to \code{FALSE} meaning that the cache in the md_branch will be used}
}
\value{
Nothing, invisibly. This is used for it's side-effect
}
\description{
(INTERNAL) Build and deploy the site with continous integration
}
\note{
this function is not for interactive use. It requires git to be
installed on your machine and will destroy anything you have in the
\verb{site/} folder.
}
\keyword{internal}
