% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_session_info.R
\name{ci_session_info}
\alias{ci_session_info}
\title{Report session information to the user}
\usage{
ci_session_info()
}
\value{
Nothing, used entirely for side-effect
}
\description{
This function is used in continuous integration to report the packages used
in building the site and pull requests
}
\keyword{internal}
