% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_schedule.R
\name{clear_episodes}
\alias{clear_episodes}
\title{Clear the schedule in the lesson}
\usage{
clear_episodes(path = ".")
}
\arguments{
\item{path}{path to the lesson}
}
\value{
NULL, invisibly
}
\description{
Clear the schedule in the lesson
}
\examples{
tmp <- tempfile()
create_lesson(tmp)
get_episodes(tmp) # produces warning
set_episodes(tmp, get_episodes(tmp), write = TRUE)
get_episodes(tmp) # no warning
clear_episodes(tmp)
get_episodes(tmp) # produces warning again because there is no schedule
}
