% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_schedule.R
\name{clear_schedule}
\alias{clear_schedule}
\title{Clear the schedule in the lesson}
\usage{
clear_schedule(path = ".")
}
\arguments{
\item{path}{path to the lesson}
}
\value{
NULL, invisibly
}
\description{
Clear the schedule in the lesson
}
\examples{
tmp <- tempfile()
create_lesson(tmp)
get_schedule(tmp) # produces warning
set_schedule(tmp, get_schedule(tmp), write = TRUE)
get_schedule(tmp) # no warning
clear_schedule(tmp)
get_schedule(tmp) # produces warning again because there is no schedule
}
