% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_site.R
\name{clear_site}
\alias{clear_site}
\title{Remove all files associated with the site}
\usage{
clear_site(path = ".")
}
\arguments{
\item{path}{the path to the site}
}
\description{
Use this if you want to rebuild your site from scratch.
}
\examples{
tmp <- tempfile()
create_lesson(tmp)
build_lesson(tmp, preview = FALSE)
dir(file.path(tmp, "site"))
clear_site(tmp)
dir(file.path(tmp, "site"))
}
