% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{copy_template}
\alias{copy_template}
\title{copy a sandpaper template file to a path with data}
\usage{
copy_template(template, path = NULL, name = NULL, values = NULL)
}
\arguments{
\item{template}{the base of a valid template function (e.g. "episode" for
\code{\link[=template_episode]{template_episode()}})}

\item{path}{the folder in which to write the file. Defaults to \code{NULL}, which
will return the filled template as a character vector}

\item{name}{the name of the file. Defaults to \code{NULL}}

\item{values}{the values to fill in the template (if any). Consult the
files in the \verb{templates/} folder of your sandpaper installation for details.}
}
\value{
a character vector if \code{path} or \code{name} is \code{NULL}, otherwise, this is
used for its side effect of creating a file.
}
\description{
copy a sandpaper template file to a path with data
}
\keyword{internal}
