% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_episode.R
\name{create_episode}
\alias{create_episode}
\alias{create_episode_md}
\alias{create_episode_rmd}
\alias{draft_episode_md}
\alias{draft_episode_rmd}
\alias{create_episode_qmd}
\alias{draft_episode_qmd}
\title{Create an Episode from a template}
\usage{
create_episode(
  title,
  ext = "Rmd",
  make_prefix = FALSE,
  add = TRUE,
  path = ".",
  open = rlang::is_interactive()
)

create_episode_md(
  title,
  make_prefix = FALSE,
  add = TRUE,
  path = ".",
  open = rlang::is_interactive()
)

create_episode_rmd(
  title,
  make_prefix = FALSE,
  add = TRUE,
  path = ".",
  open = rlang::is_interactive()
)

draft_episode_md(
  title,
  make_prefix = FALSE,
  path = ".",
  open = rlang::is_interactive()
)

draft_episode_rmd(
  title,
  make_prefix = FALSE,
  path = ".",
  open = rlang::is_interactive()
)

create_episode_qmd(
  title,
  make_prefix = FALSE,
  add = TRUE,
  path = ".",
  open = rlang::is_interactive()
)

draft_episode_qmd(
  title,
  make_prefix = FALSE,
  path = ".",
  open = rlang::is_interactive()
)
}
\arguments{
\item{title}{the title of the episode}

\item{ext}{a character. If \code{ext = "Rmd"} (default), then the new episode will
be an R Markdown episode. If \code{ext = "md"}, then the new episode will be
a markdown episode, which can not generate dynamic content. If
\code{ext = "qmd"}, then the new episode will be a Quarto document, which can
render code in multiple languages (Python, R, Julia, etc.).}

\item{make_prefix}{a logical. When \code{TRUE}, the prefix for the file will be
automatically determined by the files already present. When \code{FALSE}
(default), it assumes no prefix is needed.}

\item{add}{(logical or numeric) If numeric, it represents the position the
episode should be added. If \code{TRUE}, the episode is added to the end of the
schedule. If \code{FALSE}, the episode is added as a draft episode.}

\item{path}{the path to the \code{{sandpaper}} lesson.}

\item{open}{if interactive, the episode will open in a new editor window.}
}
\description{
These functions allow you to create an episode that will be added to the
schedule.
}
\examples{
tmp <- tempfile()
create_lesson(tmp, open = FALSE, rmd = FALSE)
create_episode_md("getting-started", path = tmp)
}
