% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sidebar.R
\name{create_resources_dropdown}
\alias{create_resources_dropdown}
\title{Create dropdown menus for extra content}
\usage{
create_resources_dropdown(files, type = "learners")
}
\arguments{
\item{files}{a vector of markdown file names}

\item{type}{one of "learners" (default) or "instructors". If it is learners,
the setup page will be excluded since it is included in the index. For
"instructors", the instructor notes are included and the learner profiles
are included.}
}
\value{
a list with character vectors of HTML list elements.
}
\description{
Create dropdown menus for extra content
}
\keyword{internal}
