% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_sidebar_item}
\alias{create_sidebar_item}
\title{Create a single item that appears in the sidebar}
\usage{
create_sidebar_item(html, name, position)
}
\description{
Create a single item that appears in the sidebar
}
