% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_syllabus.R
\name{create_syllabus}
\alias{create_syllabus}
\title{Create a syllabus for the lesson}
\usage{
create_syllabus(path = ".", questions = FALSE)
}
\arguments{
\item{path}{the path to a lesson}

\item{questions}{if \code{TRUE}, the questions in the episodes will be added to
the table. Defaults to \code{FALSE}.}
}
\value{
a data frame containing the syllabus for the lesson with the timing,
links, and questions associated
}
\description{
This function is generally for internal use, but may be useful for those who
whish to automate creation of their own home pages.
}
\keyword{internal}
