% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-fixtures.R
\name{create_test_lesson}
\alias{create_test_lesson}
\title{Create a test lesson for use in our test fixtures}
\usage{
create_test_lesson()
}
\description{
This function creates a new {sandpaper} lesson and assigns it to the
\code{sandpaper.test_fixture} global option.
}
\details{
It then returns a function that will restore the test fixture when called
with no arguments
}
