% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-store.R
\name{.lesson_store}
\alias{.lesson_store}
\title{Internal Global Lesson Storage Generator}
\usage{
.lesson_store()
}
\value{
a list with four functions that operate on a supplied path string
\itemize{
\item \code{get()} returns \code{.this_lesson} (as described in \code{\link[=this_lesson]{this_lesson()}})
\item \code{set(path)} sets \code{.this_lesson} and its git statuses from the lesson in
\code{path}. This also sets \code{set_globals()} and \code{set_resource_list()}.
\item \code{valid(path)} uses \code{path} to validate if a lesson is identical to the
stored lesson from its git status. Returns \code{TRUE} if it is identical and
\code{FALSE} if it is not
\item \code{clear()} resets all global variables.
}
}
\description{
This function will generate an object store that will contain and update a
lesson object based on the status of the git repository
}
\seealso{
\code{\link[=.list_store]{.list_store()}} for a generic list implementation and
\code{\link[=this_leson]{this_leson()}} for details of the implementation of this generator in
sandpaper
}
\keyword{internal}
