% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_workflows.R
\name{fetch_github_workflows}
\alias{fetch_github_workflows}
\title{Fetch GitHub Workflows}
\usage{
fetch_github_workflows(
  path = ".",
  files = c("comment-pr.yaml", "pr-close.yaml", "pull-request.yaml",
    "remove-branch.yaml", "sandpaper-main.yaml"),
  base = "https://raw.githubusercontent.com/zkamvar/actions/main/workflows/",
  overwrite = TRUE
)
}
\arguments{
\item{path}{path to the current lesson.}

\item{files}{a character vector of file names. Defaults to the workflows
needed for sandpaper to work on github.}

\item{base}{the base URL for the workflows}

\item{overwrite}{if \code{TRUE} (default), the file(s) will be overwritten.}
}
\value{
the output of \code{\link[usethis:use_github_action]{usethis::use_github_action()}}
}
\description{
This is a wrapper around \code{\link[usethis:use_github_action]{usethis::use_github_action()}} to download and
install github workflow files in your sandpaper lesson.
}
\note{
this assumes that you have an active internet connection.
}
