% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-translate.R
\name{fill_translation_vars}
\alias{fill_translation_vars}
\title{Apply template items to translated strings}
\usage{
fill_translation_vars(the_data)
}
\arguments{
\item{the_data}{a list of global variables (either \code{learner_globals} or
\code{instructor_globals}) that also contains a "translate" element containing
a list of translated strings.}
}
\value{
the translated list with templated data filled out
}
\description{
Apply template items to translated strings
}
\details{
There are two kinds of templating we use:
\enumerate{
\item variable templating indicated by \verb{\{\% key \%\}} where \code{key} represents a
variable that exists within the global data and is replaced.
\item link templating indicated by \verb{<(text to wrap)>} where we replace the
\verb{<()>} with a known URL or HTML markup. This allows the translators to
translate text without having to worry about HTML markup.
}
}
\examples{

dat <- list(
  a = "a barn",
  b = "a bee",
  minutes = 5,
  translate = list(
     one = "a normal translated string (pretend it's translated from another language)",
     two = "a question: are you (A) {\% a \%}, (B) {\% b \%}",
     EstimatedTime = "Estimated time: {\% icons$clock \%} {\% minutes \%}",
     license = "Licensed under {\% license \%} by {\% authors \%}",
     ThisLessonCoC = "This lesson operates under our <(Code of Conduct)>"
  )
)
asNamespace("sandpaper")$fill_translation_vars(dat)
}
\keyword{internal}
