% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-fixtures.R
\name{create_test_lesson}
\alias{create_test_lesson}
\alias{generate_restore_fixture}
\alias{setup_local_remote}
\alias{remove_local_remote}
\title{Test fixture functions for sandpaper}
\usage{
create_test_lesson()

generate_restore_fixture(tf)

setup_local_remote(
  repo,
  remote = tempfile(),
  name = "sandpaper-local",
  verbose = FALSE
)

remove_local_remote(repo, name = "sandpaper-local")
}
\arguments{
\item{tf}{(\code{generate_restore_fixture()}) the path to the test fixture lesson}

\item{repo}{path to a git repository}

\item{remote}{path to an empty or uninitialized directory. Defaults to a
tempfile}

\item{name}{of the remote, defaults to "sandpaper-local"}

\item{verbose}{if \code{TRUE}, messages and output from git will be printed to
screen. Defaults to \code{FALSE}.}
}
\value{
(\code{generate_restore_fixture}) a function that will restore the test fixture

(\code{setup_local_remote()}) the repo, invisibly

(\code{remove_local_remote()}) FALSE indicating an error or a string
indicating the path to the remote
}
\description{
These functions are for use during testing of {sandpaper} and are designed to
create a temporary lesson and associated remote repository (locally).
}
\details{
\subsection{create_test_lesson()}{

This creates the test lesson and returns a function that will restore the
test fixture when called with no arguments
}

\subsection{generate_restore_fixture()}{

This creates the restore function for the test lesson
}

\subsection{setup_local_remote()}{

Creates a local remote repository in a separate temporary folder, linked to
the fixture lesson
}

\subsection{remove_local_remote()}{

Destorys the local remote repository and removes it from the fixture lesson
}
}
\note{
These are implemented in tests/testthat/setup.md
}
\keyword{internal}
