% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-fixtures.R
\name{generate_restore_fixture}
\alias{generate_restore_fixture}
\title{Generate a function to persistantly restore a lesson test fixture}
\usage{
generate_restore_fixture(tf)
}
\arguments{
\item{tf}{the path to the test fixture lesson}
}
\value{
a function that will restore the test fixture
}
\description{
Generate a function to persistantly restore a lesson test fixture
}
\keyword{internal}
