% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_build_status}
\alias{get_build_status}
\title{Generate a data frame of markdown files to be updated}
\usage{
get_build_status(sources, built, rebuild = FALSE)
}
\arguments{
\item{built}{a vector of sandpaper-generated markdown files}

\item{rebuild}{if \code{TRUE}, all of the input files are forced to rebuild.}

\item{episodes}{a vector of full paths to RMarkdown files to be generated}
}
\description{
Get the build status for a vector of episodes against a vector of markdown
files using MD5 sums.
}
\details{
Return a list with a data frame of episodes that need to be built or
rebuilt and a vector of built episodes that need to be removed.
}
\keyword{internal}
