% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-built-db.R
\name{get_built_db}
\alias{get_built_db}
\title{Get the database of built files and their hashes}
\usage{
get_built_db(db = "site/built/md5sum.txt", filter = "*R?md")
}
\arguments{
\item{db}{the path to the database.}

\item{filter}{regex describing files to include.}
}
\value{
a data frame with three columns:
\itemize{
\item file: the path to the source file
\item checksum: the hash of the source file to generate the built file
\item built: the path to the built file
}
}
\description{
Get the database of built files and their hashes
}
\seealso{
\code{\link[=build_status]{build_status()}}, \code{\link[=get_hash]{get_hash()}}
}
\keyword{internal}
