% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Get the configuration parameters for the lesson}
\usage{
get_config(path)
}
\arguments{
\item{path}{path to the lesson}
}
\value{
a yaml list
}
\description{
Get the configuration parameters for the lesson
}
\examples{
tmp <- tempfile()
create_lesson(tmp)
get_config(tmp)
}
