% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drafts.R
\name{get_drafts}
\alias{get_drafts}
\title{Show files in draft form}
\usage{
get_drafts(
  path,
  folder = NULL,
  message = getOption("sandpaper.show_draft", TRUE)
)
}
\arguments{
\item{path}{path to the the sandpaper lesson}

\item{folder}{the specific folder for which to list the draft files. Defaults
to \code{NULL}, which indicates all folders listed in \code{config.yaml}.}

\item{message}{if \code{TRUE} (default), an informative message about the files
that are in draft status are printed to the screen.}
}
\value{
a vector of paths to files in draft and a message (if specified)
}
\description{
By default, \code{{sandpaper}} will use the files in alphabetical order as they are
presented in the folders, however, it is \strong{strongly} for authors to specify
the order of the files in their lessons, so that it's easy to rearrange or
add, split, or rearrange files.
}
\details{
This mechanism also allows authors to work on files in a draft form without
them being published. This function will list and show the files in draft for
automation and audit.
}
