% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dropdown.R
\name{get_dropdown}
\alias{get_dropdown}
\alias{get_episodes}
\alias{get_learners}
\alias{get_instructors}
\alias{get_profiles}
\title{Get the contents of a dropdown on the site}
\usage{
get_dropdown(path = ".", folder)

get_episodes(path = ".")

get_learners(path = ".")

get_instructors(path = ".")

get_profiles(path = ".")
}
\arguments{
\item{path}{the path to the lesson, defaults to the current working directory}

\item{the}{name of the folder containing the dropdown items}
}
\value{
a character vector of episodes in order of presentation
}
\description{
Get the contents of a dropdown on the site
}
\examples{
tmp <- tempfile()
create_lesson(tmp)
get_episodes(tmp)
}
