% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_episodes.R
\name{get_episodes}
\alias{get_episodes}
\title{Get the lesson schedule}
\usage{
get_episodes(path = ".")
}
\arguments{
\item{path}{the path to the lesson, defaults to the current working directory}
}
\value{
a character vector of episodes in order of presentation
}
\description{
Get the lesson schedule
}
\examples{
tmp <- tempfile()
create_lesson(tmp)
get_episodes(tmp)
}
