% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-built-db.R
\name{get_hash}
\alias{get_hash}
\title{Get the expected hash from a set of built files}
\usage{
get_hash(path, db = fs::path(path_built(path), "md5sum.txt"))
}
\arguments{
\item{path}{path to at least one generated markdown file}

\item{db}{the path to the text database}
}
\value{
a character vector of checksums
}
\description{
Get the expected hash from a set of built files
}
\seealso{
\code{\link[=build_status]{build_status()}}, \code{\link[=get_built_db]{get_built_db()}}
}
\keyword{internal}
