% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_episode.R
\name{get_nav_data}
\alias{get_nav_data}
\title{Generate the navigation data for a page}
\usage{
get_nav_data(
  path_md,
  path_src = NULL,
  home = NULL,
  this_page = NULL,
  page_back = NULL,
  page_forward = NULL
)
}
\arguments{
\item{path_md}{the path to the episode markdown (not RMarkdown) file
(usually via \code{\link[=build_episode_md]{build_episode_md()}}).}

\item{path_src}{the default is \code{NULL} indicating that the source file should
be determined from the \code{sandpaper-source} entry in the yaml header. If this
is not present, then this option allows you to specify that file.}

\item{home}{the path to the lesson home}

\item{this_page}{the current page relative html address}

\item{page_back}{the URL for the previous page}

\item{page_forward}{the URL for the next page}
}
\description{
Generate the navigation data for a page
}
\keyword{internal}
