% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-paths-source.R
\name{get_resource_list}
\alias{get_resource_list}
\title{Get the full resource list of markdown files}
\usage{
get_resource_list(path, trim = FALSE)
}
\arguments{
\item{path}{path to the lesson}

\item{trim}{if \code{TRUE}, trim the paths to be relative to the lesson directory.
Defaults to \code{FALSE}, which will return the absolute paths}
}
\value{
a list of files by subfolder in the order they should appear in the
menu.
}
\description{
Get the full resource list of markdown files
}
\seealso{
\code{\link[=build_status]{build_status()}}
}
\keyword{internal}
