% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_syllabus.R
\name{get_syllabus}
\alias{get_syllabus}
\title{Create a syllabus for the lesson}
\usage{
get_syllabus(path = ".", questions = FALSE, use_built = TRUE)
}
\arguments{
\item{path}{the path to a lesson}

\item{questions}{if \code{TRUE}, the questions in the episodes will be added to
the table. Defaults to \code{FALSE}.}

\item{use_built}{if \code{TRUE} (default), the rendered episodes will be used to
generate the syllabus}
}
\value{
a data frame containing the syllabus for the lesson with the timing,
links, and questions associated
}
\description{
This function is generally for internal use, but may be useful for those who
whish to automate creation of their own home pages.
}
\keyword{internal}
