% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_this_lesson}
\alias{set_this_lesson}
\alias{clear_this_lesson}
\alias{this_lesson}
\title{Internal cache for storing lesson objects between operations inside a running
{sandpaper} session.}
\usage{
set_this_lesson(path)

clear_this_lesson()

this_lesson(path)
}
\arguments{
\item{path}{a path to the current lesson}
}
\description{
Internal cache for storing lesson objects between operations inside a running
{sandpaper} session.
}
\details{
\code{this_lesson()} will return a \link[pegboard:Lesson]{pegboard::Lesson} object if it has
previously been stored and the path is equal to that of the one stored.
Otherwise, it returns NULL.
}
\keyword{internal}
