% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_images.R
\name{make_images_section}
\alias{make_images_section}
\title{Make a section of aggregated images}
\usage{
make_images_section(name, contents, parent)
}
\arguments{
\item{name}{the name of the section, (may or may not be prefixed with \verb{images-})}

\item{contents}{an \code{xml_nodeset} of figure elements from \code{\link[=get_content]{get_content()}}}

\item{parent}{the parent div of the images page}
}
\value{
the section that was added to the parent
}
\description{
This will insert xml figure nodes into the images page, printing the alt text
descriptions for users who are not using screen readers.
}
\examples{
if (FALSE) {
  lsn <- "/path/to/lesson"
  pkg <- pkgdown::as_pkgdown(fs::path(lsn, "site"))

  # read in the All in One page and extract its content
  img <- get_content("images", content = "self::*", pkg = pkg)
  fig_content <- get_content("01-introduction", content = "/figure", pkg = pkg)
  make_images_section("01-introduction", contents = fig_content, parent = img)
}
}
\seealso{
\code{\link[=build_images]{build_images()}}, \code{\link[=get_content]{get_content()}}
}
\keyword{internal}
