% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_instructor_notes.R
\name{make_instructornotes_section}
\alias{make_instructornotes_section}
\title{Make a section of aggregated instructor notes}
\usage{
make_instructornotes_section(name, contents, parent)
}
\arguments{
\item{name}{the name of the section, (may or may not be prefixed with \verb{images-})}

\item{contents}{an \code{xml_nodeset} of figure elements from \code{\link[=get_content]{get_content()}}}

\item{parent}{the parent div of the images page}
}
\value{
the section that was added to the parent
}
\description{
This will append instructor notes from the inline sections of the lesson to
the instructor-notes page, separated by section and \verb{<hr>} elements.
}
\note{
On the learner view, instructor notes will not be present
}
\examples{
if (FALSE) {
lsn <- "/path/to/lesson"
pkg <- pkgdown::as_pkgdown(fs::path(lsn, "site"))

# read in the All in One page and extract its content
notes <- get_content("instructor-notes", content =
  "section[@id='aggregate-instructor-notes']", pkg = pkg, instructor = TRUE)
agg <- "/div[contains(@class, 'instructor-note')]//div[@class='accordion-body']"
note_content <- get_content("01-introduction", content = agg, pkg = pkg)
make_instructornotes_section("01-introduction", contents = note_content, 
  parent = notes)

# NOTE: if the object for "contents" ends with "_learn", no content will be
# appended
note_learn <- note_content
make_instructornotes_section("01-introduction", contents = note_learn, 
  parent = notes)

}
}
\seealso{
\code{\link[=build_instructor_notes]{build_instructor_notes()}}, \code{\link[=get_content]{get_content()}}
}
\keyword{internal}
