% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_python.R
\name{py_install}
\alias{py_install}
\title{Install Python packages and add them to the cache}
\usage{
py_install(path = ".", packages, ...)
}
\arguments{
\item{path}{path to your lesson. Defaults to the current working directory.}

\item{packages}{Python packages to be installed as a character vecto.}

\item{...}{Further arguments to be passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}}
}
\description{
Installs Python packages with \code{\link[reticulate:py_install]{reticulate::py_install()}} and then records
them in the renv environment. This ensures \code{\link[=manage_deps]{manage_deps()}} keeps track of the
Python packages as well.
}
\details{
Unlike with R packages, \pkg{renv} is not yet capable of automatically
detecting Python dependencies. Therefore, this helper function is provided to
correctly install Python packages and recording them in the renv environment.
Subsequent calls of \code{\link[=manage_deps]{manage_deps()}} will then correctly restore the required
Python packages if needed.
}
