% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-aggregate.R
\name{read_all_html}
\alias{read_all_html}
\title{read all HTML files in a folder}
\usage{
read_all_html(path)
}
\arguments{
\item{path}{the path to a folder with HTML files}
}
\value{
a nested list of \code{html_documents} from \code{\link[xml2:read_xml]{xml2::read_html()}} with two
top-level elements:
\itemize{
\item \verb{$learner}: contains all of the html pages in the learner view
\item \verb{$instructor}: all of the pages in the instructor view
\item \verb{$paths}: the absolute paths for the pages
}
}
\description{
read all HTML files in a folder
}
\examples{
tmpdir <- tempfile()
on.exit(fs::dir_delete(tmpdir))
fs::dir_create(tmpdir)
fs::dir_create(fs::path(tmpdir, "instructor"))
writeLines("<p>Instructor</p>", fs::path(tmpdir, "instructor", "index.html"))
writeLines("<p>Learner</p>", fs::path(tmpdir, "index.html"))
sandpaper:::read_all_html(tmpdir)

}
