% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_pdf.R
\name{render_pdf}
\alias{render_pdf}
\title{Render pdf from a markdown file}
\usage{
render_pdf(path_in, ..., quiet = FALSE)
}
\arguments{
\item{path_in}{path to a markdown file}

\item{...}{extra options (e.g. lua filters) to be passed to pandoc}

\item{quiet}{if \code{TRUE}, no output is produced. Default is \code{FALSE}, which
reports the markdown build via pandoc}
}
\description{
This uses \code{\link[rmarkdown:pandoc_convert]{rmarkdown::pandoc_convert()}} to render PDF from a markdown file.
We've specified pandoc extensions that align with the features desired in the
Carpentries
}
\keyword{internal}
