% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renv.R
\name{renv_diagnostics}
\alias{renv_diagnostics}
\title{Print a diagnostics report for the package cache}
\usage{
renv_diagnostics(path = ".", profile = "lesson-requirements")
}
\arguments{
\item{path}{the path to the lesson to use for diagnostics}

\item{profile}{the profile to work with (defaults to "lesson-requirements"}
}
\description{
Print a diagnostics report for the package cache
}
\keyword{internal}
