% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renv.R
\name{renv_lockfile_hash}
\alias{renv_lockfile_hash}
\title{Get the hash for the previous and current lockfile (as recorded in the lesson)}
\usage{
renv_lockfile_hash(path, db_path, profile = "lesson-requirements")
}
\arguments{
\item{path}{path to the lesson}

\item{db_path}{path to the database}

\item{profile}{name of the profile renv uses for the lesson requirements}
}
\value{
a named list:
\itemize{
\item old: hash value recoreded in the database
\item new: hash value of the current file
}
}
\description{
Get the hash for the previous and current lockfile (as recorded in the lesson)
}
\keyword{internal}
