% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-built-db.R
\name{reserved_db}
\alias{reserved_db}
\title{Filter reserved files from the built db}
\usage{
reserved_db(db)
}
\arguments{
\item{db}{the database from \code{\link[=get_built_db]{get_built_db()}}}
}
\value{
a data frame, but a bit shorter
}
\description{
Filter reserved files from the built db
}
\seealso{
\code{\link[=get_built_db]{get_built_db()}}
}
\keyword{internal}
