% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_episodes.R
\name{reset_episodes}
\alias{reset_episodes}
\title{Clear the schedule in the lesson}
\usage{
reset_episodes(path = ".")
}
\arguments{
\item{path}{path to the lesson}
}
\value{
NULL, invisibly
}
\description{
Clear the schedule in the lesson
}
\examples{
tmp <- tempfile()
create_lesson(tmp)
get_episodes(tmp) # produces warning
set_episodes(tmp, get_episodes(tmp), write = TRUE)
get_episodes(tmp) # no warning
reset_episodes(tmp)
get_episodes(tmp) # produces warning again because there is no schedule
}
