% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_site.R
\name{reset_site}
\alias{reset_site}
\title{Remove all files associated with the site}
\usage{
reset_site(path = ".")
}
\arguments{
\item{path}{the path to the site}
}
\description{
Use this if you want to rebuild your site from scratch.
}
\examples{
tmp <- tempfile()
create_lesson(tmp)
if (rmarkdown::pandoc_available("2.11"))
  build_lesson(tmp, preview = FALSE)
dir(file.path(tmp, "site"))
reset_site(tmp)
dir(file.path(tmp, "site"))
}
