% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{sandpaper.options}
\alias{sandpaper.options}
\title{Global Options}
\description{
this is some documentation about options
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{option("sandpaper.show_draft" = TRUE)
option("sandpaper.links" = NULL)
option("sandpaper.use_renv" = FALSE)
option("sandpaper.package_cache_trigger" = FALSE)
option("sandpaper.test_fixture" = NULL)
}\if{html}{\out{</div>}}

As of 2022-02-22, there are several options that are used in sandpaper that
may be manipulated by the user. This set may change in the future, but here
are the description of these options and how they are set on startup:
\subsection{sandpaper.show_draft}{

\strong{Default: \code{TRUE}} This is for user messages. If \code{TRUE}, a message about
episodes in draft status (i.e. episodes that are in the folder, but not in
the schedule) will be printed with \code{get_drafts()}. Setting this option to
\code{FALSE} will turn off this feature.
}

\subsection{sandpaper.links}{

\strong{Default: \code{NULL}} This option provides a way to override the default place
for links in your sandpaper lesson. If it is NULL and there is a file called
\code{links.md} at the top of the repository, this will be appended to the bottom
of each page before it is rendered to HTML.
}

\subsection{sandpaper.use_renv}{

\strong{Default: variable} This option should not be modified by the user. It
determines if \code{{renv}} should be used locally for R-based lessons. It is set
by \code{\link[=use_package_cache]{use_package_cache()}} and unset by \code{\link[=no_package_cache]{no_package_cache()}}. If a local user
has never consented to using \code{{renv}} previously, then it defaults to \code{FALSE},
but if \code{{renv}} has previously been used, it will be \code{TRUE}.
}

\subsection{sandpaper.package_cache_trigger}{

\strong{Default: FALSE locally/TRUE on GitHub} this tells R Markdown lessons to
rebuild everything if the \code{{renv}} lockfile changes.
}

\subsection{sandpaper.test_fixture}{

\strong{Default: NULL} This is ONLY for internal use for testing interactive
components non-interactively and for setting \code{{renv}} to behave correctly while
testing.
}
}
