% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandpaper_site.R
\name{sandpaper_site}
\alias{sandpaper_site}
\title{Site generator for sandpaper}
\usage{
sandpaper_site(input = ".", ...)
}
\description{
This is a custom site generator for compatibility with
\code{\link[rmarkdown:render_site]{rmarkdown::site_generator()}}. For RStudio users, \strong{placing this in
the \code{index.md} yaml header will make the knit button work}:
}
\details{
\if{html}{\out{<div class="yaml">}}\preformatted{site: sandpaper::sandpaper_site
}\if{html}{\out{</div>}}

This will be automatically added to the \code{index.md} for all sandpaper sites
from version 0.0.0.9013, so maintainers should not need to worry about this.

Thanks goes to Yihui Xie for coming up with the concept of modular site
generators and for JJ Alaire for allowing sub-directory structures for
RMarkdown sites.
}
\keyword{internal}
