% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dropdown.R
\name{set_dropdown}
\alias{set_dropdown}
\alias{set_config}
\alias{set_episodes}
\alias{set_learners}
\alias{set_instructors}
\alias{set_profiles}
\title{Set the order of items in a dropdown menu}
\usage{
set_dropdown(path = ".", order = NULL, write = FALSE, folder)

set_config(key = NULL, value = NULL, path = ".", write = FALSE)

set_episodes(path = ".", order = NULL, write = FALSE)

set_learners(path = ".", order = NULL, write = FALSE)

set_instructors(path = ".", order = NULL, write = FALSE)

set_profiles(path = ".", order = NULL, write = FALSE)
}
\arguments{
\item{path}{path to the lesson. Defaults to the current directory.}

\item{order}{the files in the order presented (with extension)}

\item{write}{if \code{TRUE}, the schedule will overwrite the schedule in the
current file.}

\item{folder}{one of four folders that sandpaper recognises where the files
listed in \code{order} are located: episodes, learners, instructors, profiles.}

\item{key}{the key for a vector of parameters}

\item{value}{the value matching each key}
}
\description{
Set the order of items in a dropdown menu
}
\examples{

tmp <- tempfile()
create_lesson(tmp, "test lesson")
# Change the title and License
set_config(key = c("title", "license"), 
  value = c("Absolutely Free Lesson", "CC0"),
  path = tmp,
  write = TRUE
)
create_episode("using-R", path = tmp)
print(sched <- get_episodes(tmp))

# reverse the schedule
set_episodes(tmp, order = rev(sched))
# write it
set_episodes(tmp, order = rev(sched), write = TRUE)

# see it
get_episodes(tmp)

}
