% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-varnish.R
\name{set_globals}
\alias{set_globals}
\title{Set the necessary common global variables for use in the {varnish} template.}
\usage{
set_globals(path)
}
\arguments{
\item{path}{the path to the lesson}
}
\description{
This will enforce three global lists:
}
\details{
\enumerate{
\item \code{.resources}, which is equivalent to the output of \code{get_source_list()}
\item \code{learner_globals} the navigation items for the learners
\item \code{instructor_globals} the namvigation items for the instructors
}

The things that are added:
\itemize{
\item \code{sidebar} This is generated from \code{\link[=create_sidebar]{create_sidebar()}} and is the same in the
learner and instructor globals
\item \code{more} This is the "More" dorpdown menu, which is created via \code{\link[=create_resources_dropdown]{create_resources_dropdown()}}.
\item \code{resources} The same as "More", but positioned on the mobile sidebar.
\item \verb{\{sandpaper,varnish,pegboard\}_version} package versions of each package.
}
}
\keyword{internal}
