% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_schedule.R
\name{set_schedule}
\alias{set_schedule}
\title{Set the schedule for the lesson}
\usage{
set_schedule(path, order = NULL, write = FALSE)
}
\arguments{
\item{path}{path to the lesson}

\item{order}{the episodes in the order presented (with extension)}

\item{write}{if \code{TRUE}, the schedule will overwrite the schedule in the
current file.}
}
\description{
Set the schedule for the lesson
}
\examples{

tmp <- tempfile()
create_lesson(tmp)
create_episode("using-R", path = tmp)
print(sched <- get_schedule(tmp))

# reverse the schedule
set_schedule(tmp, order = rev(sched))
# write it
set_schedule(tmp, order = rev(sched), write = TRUE)

# see it
get_schedule(tmp)

}
