% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-git.R
\name{setup_local_remote}
\alias{setup_local_remote}
\alias{remove_local_remote}
\title{Helper functions for tests to set a local remote called "sandpaper-local"}
\usage{
setup_local_remote(
  repo,
  remote = tempfile(),
  name = "sandpaper-local",
  verbose = FALSE
)

remove_local_remote(name = "sandpaper-local", repo)
}
\arguments{
\item{repo}{path to a git repository}

\item{remote}{path to an empty or uninitialized directory. Defaults to a
tempfile}

\item{name}{of the remote, defaults to "sandpaper-local"}

\item{verbose}{if \code{TRUE}, messages and output from git will be printed to
screen. Defaults to \code{FALSE}.}
}
\value{
repo, invisibly
}
\description{
Helper functions for tests to set a local remote called "sandpaper-local"
}
\keyword{internal}
