% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_episode.R
\name{strip_prefix}
\alias{strip_prefix}
\title{This will strip existing episode prefixes and set the schedule}
\usage{
strip_prefix(path = ".", write = FALSE)
}
\arguments{
\item{path}{the path to the lesson (defaults to the current working directory)}

\item{write}{defaults to \code{FALSE}, which will show the potential changes.
If \code{TRUE}, the schedule will be modified and written to \code{config.yaml}}
}
\value{
when \code{write = TRUE}, the modified list of episodes. When
\code{write = FALSE}, the modified call is returned.
}
\description{
Episode order for Carpentries lessons originally used a strategy of prefixing
files by a two-digit number to force a specific order by filename. This
function will strip these numbers from the filename and set the schedule
according to the original order.
}
\note{
git will recognise this as deleting a file and then adding a new file
in the stage. If you run \verb{git add}, it should recognise that it is a rename.
}
\examples{
if (FALSE) {
  strip_prefix() # test if the function is doing what you want it to do
  strip_prefix(write = TRUE) # rewrite the episode names
}
}
\seealso{
\code{\link[=create_episode]{create_episode()}} for creating new episodes, \code{\link[=move_episode]{move_episode()}} for
moving individual episodes around.
}
