% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_workflows.R
\name{update_github_workflows}
\alias{update_github_workflows}
\title{Update github workflows}
\usage{
update_github_workflows(path = ".", files = "", overwrite = TRUE)
}
\arguments{
\item{path}{path to the current lesson.}

\item{files}{the files to include in the update. Defaults to an empty string,
which will update all files}

\item{overwrite}{if \code{TRUE} (default), the file(s) will be overwritten.}
}
\value{
the paths to the new files.
}
\description{
This function copies and updates the workflows to run {sandpaper}.
}
\note{
this assumes that you have an active internet connection.
}
