% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_schedule.R
\name{update_schedule}
\alias{update_schedule}
\title{Create the episode schedule for the lesson}
\usage{
update_schedule(path)
}
\arguments{
\item{path}{path to your lesson}

\item{write}{if \code{TRUE}, the modified schedule will be written to your config
file. Defaults to \code{FALSE}.}
}
\value{
NULL, invisibly. This is used for the side-effect of modifying the
\code{config.yml} file.
}
\description{
Note: this is not a public-facing function
}
\details{
This will create a schedule in the \code{config.yml} file for the episodes in the
\verb{episodes/} directory. By default, it will use alphabetical order, but you
can rearrange the episodes in the \code{config.yml} file manually. The \code{config.yml}
file will be the source of truth for building the site, so creating the
schedule in that file is important.
}
\keyword{internal}
