% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_varnish.R
\name{update_varnish}
\alias{update_varnish}
\title{Update the local version of the carpentries style}
\usage{
update_varnish(version = NULL, ...)
}
\arguments{
\item{version}{if NULL, update the latest version, otherwise, this can be a
version string to identify the specific version to use.}

\item{...}{arguments passed on to \code{\link[utils:install.packages]{utils::install.packages()}}}
}
\value{
NULL, invisibly
}
\description{
Update the local version of the carpentries style
}
\note{
this requires an internet connection
}
