% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_deps.R
\name{use_package_cache}
\alias{use_package_cache}
\title{Give Consent to Use Package Cache}
\usage{
use_package_cache(prompt = interactive(), quiet = prompt)
}
\value{
nothing. this is used for its side-effect
}
\description{
\subsection{Summary}{
}
}
\details{
This function explicitly gives \pkg{sandpaper} permission to use \pkg{renv}
to create a package cache for this and future lessons.
\subsection{Background}{

By default, \pkg{sandpaper} will happily build your lesson using the packages
available in your default R library, but this can be undesirable for a couple
of reasons:
\enumerate{
\item You may have a different version of a lesson package that is used on the
lesson website, which may result in strange errors, warnings, or incorrect
output.
\item You might be very cautious about updating any components of your current
R infrastructure because your work depends on you having the correct
package versions installed.
}

To alleviate these concerns, \pkg{sandpaper} uses the \pkg{renv} package to
generate a lesson-specific library that has package versions pinned until the
lesson authors choose to update them. This is designed to be
minimally-invasive, using the packages you already have and downloading from
external repositories only when necessary.
}

\subsection{What if I've used \pkg{renv} before?}{

If you have used \pkg{renv} in the past, then there is no need to give
consent to use the cache.
}
}
