% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_python.R
\name{use_python}
\alias{use_python}
\title{Use Python}
\usage{
use_python(
  path = ".",
  python = NULL,
  type = c("auto", "virtualenv", "conda", "system"),
  ...
)
}
\arguments{
\item{path}{path to the current project}

\item{python}{The path to the version of Python to be used with this project. See
\strong{Finding Python} for more details.}

\item{type}{The type of Python environment to use. When \code{"auto"} (the default),
virtual environments will be used.}

\item{...}{Further arguments to be passed on to \code{\link[renv:use_python]{renv::use_python()}}}
}
\value{
The path to the Python executable. Note that this function is mainly
called for its side effects.
}
\description{
Associate a version of Python with your lesson. This is essentially a wrapper
around \code{\link[renv:use_python]{renv::use_python()}}. To add Python packages, use \code{\link[=py_install]{py_install()}}.
}
\seealso{
\code{\link[renv:use_python]{renv::use_python()}}
}
