% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_lesson.R
\name{validate_lesson}
\alias{validate_lesson}
\title{Pre-build validation of lesson elements}
\usage{
validate_lesson(path = ".", headings = FALSE, quiet = FALSE)
}
\arguments{
\item{path}{the path to the lesson. Defaults ot the current directory}

\item{headings}{If \code{TRUE}, headings will be checked and validated. Currently
set to \code{FALSE} as we are re-investigating some false positives.}

\item{quiet}{if \code{TRUE}, no messages will be issued, otherwise progress
messages will be issued for each test}
}
\value{
a list with the results for each test as described in
\link[pegboard:Lesson]{pegboard::Lesson}
}
\description{
A validator based on the \link[pegboard:Lesson]{pegboard::Lesson} class cached with \code{\link[=this_lesson]{this_lesson()}}
that will provide line reports for fenced divs, links, images, and heading
structure. For details on the type of validators avaliable, see the
\code{{pegboard}} article \href{https://carpentries.github.io/pegboard/articles/validation.html}{Validation of Lesson Elements}
}
\details{
\subsection{Headings}{

We expect the headings to be semantic and informative. Details of the tests
for headings can be found at \code{\link[pegboard:validate_headings]{pegboard::validate_headings()}}
}

\subsection{Internal Links and Images}{

Internal links and images should exist and images should have alt text.
Details for these tests can be found at \code{\link[pegboard:validate_links]{pegboard::validate_links()}}
}

\subsection{Fenced Divs (callout blocks)}{

Callout Blocks should be one of the expected types. Details for this test
can be found at \code{\link[pegboard:validate_divs]{pegboard::validate_divs()}}
}
}
\examples{
tmp <- tempfile()
lsn <- create_lesson(tmp, open = FALSE)
validate_lesson(lsn, headings = TRUE)
}
\keyword{internal}
