% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-renv.R
\name{with_renv_factory}
\alias{with_renv_factory}
\title{Generate a function to run in a renv profile}
\usage{
with_renv_factory(func, renv_path, renv_profile = "lesson-requirements")
}
\arguments{
\item{func}{The function to be evaluated after loading the renv environment.}

\item{renv_path}{The path to the renv environment to load. Usually a directory created by
\code{\link[=create_lesson]{create_lesson()}}}

\item{renv_profile}{Optional profile to load. Defaults to "lesson-requirements".}

\item{...}{Additional arguments to be passed to \code{func}.}
}
\value{
The result of evaluating \code{func(...)} after loading the renv environment.
}
\description{
This is a \href{https://adv-r.hadley.nz/function-operators.html}{Function operator} which will
generate a function that will run in a renv profile. This is useful for running code in a
separate R subprocess with \code{\link[callr:r]{callr::r()}}, to avoid \emph{renv} side effects related to interactive
sessions.
}
\keyword{internal}
