% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_aio.R
\name{build_aio}
\alias{build_aio}
\title{Build the All-in-one page}
\usage{
build_aio(pkg, quiet)
}
\arguments{
\item{pkg}{an object created by {pkgdown}, supplied by \code{\link[=build_site]{build_site()}}.}

\item{quiet}{If \code{TRUE} then no messages will be shown when building.

This function will build the all-in-one page for the lesson website. Because
the bottleneck is often the internal processes of {pkgdown}, we are first
templating the AIO page as a blank page and then adding in the contents using
{xml2}.}
}
\description{
Build the All-in-one page
}
\keyword{internal}
