% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_citation.R
\name{build_citation}
\alias{build_citation}
\title{Create a citation page for a lesson from a CITATION.CFF file}
\usage{
build_citation(pkg, quiet = FALSE)
}
\arguments{
\item{path}{the path to the lesson. Defaults to current working directory}

\item{out}{the path to the citation page. When this is \code{NULL} (default)
or \code{TRUE}, the output will be \code{site/built/citation.html}.}
}
\description{
Create a citation page for a lesson from a CITATION.CFF file
}
\keyword{internal}
