% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lesson.R
\name{build_lesson}
\alias{build_lesson}
\title{Build your lesson site}
\usage{
build_lesson(
  path = ".",
  rebuild = FALSE,
  quiet = !interactive(),
  preview = TRUE,
  override = list()
)
}
\arguments{
\item{path}{the path to your repository (defaults to your current working
directory)}

\item{rebuild}{if \code{TRUE}, everything will be built from scratch as if there
was no cache. Defaults to \code{FALSE}, which will only build markdown files that
haven't been built before.}

\item{quiet}{when \code{TRUE}, output is supressed}

\item{preview}{if \code{TRUE}, the rendered website is opened in a new window}

\item{override}{options to override (e.g. building to alternative paths).
This is used internally and will likely be changed.}
}
\value{
\code{TRUE} if it was successful, a character vector of issues if it was
unsuccessful.
}
\description{
This function orchestrates rendering generated lesson content and applying
the theme for the HTML site.
}
\examples{

tmp <- tempfile()
create_lesson(tmp, open = FALSE)
create_episode("first-script", path = tmp)
check_lesson(tmp)
if (rmarkdown::pandoc_available())
  build_lesson(tmp)
}
