% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_markdown.R
\name{build_markdown}
\alias{build_markdown}
\title{Build plain markdown from the RMarkdown episodes}
\usage{
build_markdown(path = ".", rebuild = FALSE, quiet = FALSE, slug = NULL)
}
\arguments{
\item{path}{the path to your repository (defaults to your current working
directory)}

\item{rebuild}{if \code{TRUE}, everything will be built from scratch as if there
was no cache. Defaults to \code{FALSE}, which will only build markdown files that
haven't been built before.}
}
\value{
\code{TRUE} if it was successful, a character vector of issues if it was
unsuccessful.
}
\description{
In the spirit of \code{{hugodown}}, This function will build plain markdown files
as a minimal R package in the \verb{site/} folder of your \code{{sandpaper}} lesson
repository tagged with the hash of your file to ensure that only files that
have changed are rebuilt.
}
\seealso{
\code{\link[=build_episode_md]{build_episode_md()}}
}
\keyword{internal}
