% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_site.R
\name{build_site}
\alias{build_site}
\title{Wrapper for site builder}
\usage{
build_site(
  path = ".",
  quiet = !interactive(),
  preview = TRUE,
  override = list(),
  slug = NULL,
  built = NULL
)
}
\arguments{
\item{path}{the path to your repository (defaults to your current working
directory)}

\item{quiet}{when \code{TRUE}, output is supressed}

\item{preview}{if \code{TRUE}, the rendered website is opened in a new window}

\item{override}{options to override (e.g. building to alternative paths).
This is used internally and will likely be changed.}

\item{slug}{The slug for the file to preview in RStudio.
If this is \code{NULL}, the preview will default to the home page. If you have
an episode whose slug is 01-introduction, then setting \code{slug = "01-introduction"} will allow RStudio to open the preview window to the
right page.}

\item{built}{a character vector of newly built files or NULL.}
}
\description{
Wrapper for site builder
}
\note{
this assumes that the markdown files have already been built and will
not work otherwise
}
\keyword{internal}
