% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-built-db.R
\name{build_status}
\alias{build_status}
\title{Identify what files need to be rebuilt}
\usage{
build_status(
  sources,
  db = "site/built/md5sum.txt",
  rebuild = FALSE,
  write = FALSE
)
}
\arguments{
\item{sources}{a character vector of ALL source files.}

\item{db}{the path to the database}

\item{rebuild}{if the files should be rebuilt, set this to TRUE (defaults to
FALSE)}

\item{write}{if TRUE, the database will be updated, Defaults to FALSE, meaning that the database will remain the same.}
}
\description{
Identify what files need to be rebuilt
}
\seealso{
\code{\link[=get_resource_list]{get_resource_list()}}, \code{\link[=get_built_db]{get_built_db()}}, \code{\link[=get_hash]{get_hash()}}
}
\keyword{internal}
