% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lesson.R
\name{check_lesson}
\alias{check_lesson}
\title{\link{deprecated} Check the lesson structure for errors}
\usage{
check_lesson(path = ".")
}
\arguments{
\item{path}{the path to your lesson}
\item{quiet}{if quiet (default \code{TRUE}) then no info messages printed to stdout}
}
\value{
\code{TRUE} (invisibly) if the lesson is cromulent, otherwise, it will
error with a list of things to fix.
}
\description{
This function is now deprecated in favour of \code{\link[=validate_lesson]{validate_lesson()}}.
}
\examples{

# Everything should work out of the box
tmp <- tempfile()
create_lesson(tmp, open = FALSE, rmd = FALSE)
check_lesson(tmp)

# if things do not work, then an error is thrown with information about
# what has failed you
unlink(file.path(tmp, ".gitignore"))
unlink(file.path(tmp, "site"), recursive = TRUE)
try(check_lesson(tmp, quiet = FALSE))

unlink(tmp)
}
\keyword{internal}
