% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lesson.R
\name{check_lesson}
\alias{check_lesson}
\title{Check the lesson structure for errors}
\usage{
check_lesson(path = ".")
}
\arguments{
\item{path}{the path to your lesson}
}
\value{
\code{TRUE} (invisibly) if the lesson is cromulent, otherwise, it will
error with a list of things to fix.
}
\description{
Check the lesson structure for errors
}
\examples{

# Everything should work out of the box
tmp <- tempfile()
create_lesson(tmp)
check_lesson(tmp)

# if things do not work, then an error is thrown with information about 
# what has failed you
unlink(file.path(tmp, ".gitignore"))
unlink(file.path(tmp, "site"), recursive = TRUE)
try(check_lesson(tmp))

unlink(tmp)
}
