% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_episode.R
\name{create_episode}
\alias{create_episode}
\title{Create an Episode from a template}
\usage{
create_episode(title, make_prefix = TRUE, path = ".")
}
\arguments{
\item{title}{the title of the episode}

\item{make_prefix}{a logical. If \code{TRUE}, the prefix for the file will be
automatically determined by the files already present. Otherwise, it assumes
you have added the prefix.}

\item{path}{the path to the {sandpaper} lesson.}
}
\description{
Create an Episode from a template
}
\examples{
tmp <- tempfile()
create_lesson(tmp)
create_episode("getting-started", path = tmp)
}
