% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lesson.R
\name{create_lesson}
\alias{create_lesson}
\title{Create a carpentries lesson}
\usage{
create_lesson(
  path,
  name = fs::path_file(path),
  rmd = TRUE,
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive(),
  add_python = FALSE,
  python = NULL,
  type = c("auto", "virtualenv", "conda", "system")
)
}
\arguments{
\item{path}{the path to the new lesson folder}

\item{name}{the name of the lesson. If not provided, the folder name will be used.}

\item{rmd}{logical indicator if the lesson should use R Markdown (\code{TRUE},
default), or if it should use Markdown (\code{FALSE}). Note that lessons can be
converted to use R Markdown at any time by adding a file with the \code{.Rmd}
file extension in the lesson.}

\item{rstudio}{create an RStudio project (defaults to if RStudio exits)}

\item{open}{if interactive, the lesson will open in a new editor window.}

\item{add_python}{if set to \code{TRUE}, will add Python as a dependency for the
lesson. See \code{\link[=use_python]{use_python()}} for details. Defaults to \code{FALSE}.}

\item{python}{the path to the version of Python to be used. The default,
\code{NULL}, will prompt the user to select an appropriate version of Python in
interactive sessions. In non-interactive sessions, \pkg{renv} will attempt
to automatically select an appropriate version. See \code{\link[renv:use_python]{renv::use_python()}}
for more details.}

\item{type}{the type of Python environment to use. When \code{"auto"}, the
default, virtual environments will be used. See \code{\link[renv:use_python]{renv::use_python()}} for
more details.}
}
\value{
the path to the new lesson
}
\description{
This will create a boilerplate directory structure for a Carpentries lesson
and initialize a git repository.
}
\examples{
tmp <- tempfile()
on.exit(unlink(tmp))
lsn <- create_lesson(tmp, name = "This Lesson", open = FALSE)
lsn
}
