% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lesson.R
\name{create_lesson}
\alias{create_lesson}
\title{Create a carpentries lesson}
\usage{
create_lesson(
  path,
  name = fs::path_file(path),
  rstudio = rstudioapi::isAvailable(),
  open = rlang::is_interactive()
)
}
\arguments{
\item{path}{the path to the new lesson folder}

\item{name}{the name of the lesson. If not provided, the folder name will be used.}

\item{rstudio}{create an RStudio project (defaults to if RStudio exits)}

\item{open}{if interactive, the lesson will open in a new editor window.}
}
\value{
the path to the new lesson
}
\description{
This will create a boilerplate directory structure for a Carpentries lesson
and initialize a git repository.
}
\examples{
tmp <- tempfile()
on.exit(unlink(tmp))
lsn <- create_lesson(tmp, name = "This Lesson", open = FALSE)
lsn
}
