% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_workflows.R
\name{fetch_github_workflows}
\alias{fetch_github_workflows}
\title{Fetch GitHub Workflows}
\usage{
fetch_github_workflows(
  files = c("comment-pr.yaml", "pr-close.yaml", "pull-request.yaml",
    "remove-branch.yaml", "sandpaper-main.yaml"),
  base = "https://raw.githubusercontent.com/zkamvar/actions/main/workflows/"
)
}
\arguments{
\item{files}{a character vector of file names}

\item{base}{the base URL for the workflows}
}
\value{
the output of \code{\link[usethis:use_github_action]{usethis::use_github_action()}}
}
\description{
This is a wrapper around \code{\link[usethis:use_github_action]{usethis::use_github_action()}} to download and
install github workflow files in your sandpaper lesson.
}
\note{
this assumes that you have an active internet connection.
}
