% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_site.R
\name{reset_site}
\alias{reset_site}
\title{Remove all files associated with the site}
\usage{
reset_site(path = ".")
}
\arguments{
\item{path}{the path to the site}
}
\description{
Use this if you want to rebuild your site from scratch.
}
\examples{
\dontshow{if (sandpaper:::example_can_run()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tmp <- tempfile()
create_lesson(tmp, open = FALSE, rmd = FALSE)
build_lesson(tmp, preview = FALSE)
dir(file.path(tmp, "site"))
reset_site(tmp)
dir(file.path(tmp, "site"))
\dontshow{\}) # examplesIf}
}
