% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandpaper-package.R
\docType{package}
\name{sandpaper-package}
\alias{sandpaper}
\alias{sandpaper-package}
\title{sandpaper: Create and Curate Carpentries Lessons}
\description{
Boilerplate directory structure, build tools, and hosting tools for your lesson.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/carpentries/sandpaper/}
  \item Report bugs at \url{https://github.com/carpentries/sandpaper/issues/}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

}
\keyword{internal}
