% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandpaper-package.R
\docType{package}
\name{sandpaper-package}
\alias{sandpaper}
\alias{sandpaper-package}
\title{sandpaper: Create and Curate Carpentries Lessons}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

We provide tools to build a Carpentries-themed lesson repository into an accessible standalone static website. These include local tools and those designed to be used in a continuous integration context so that all the lesson author needs to focus on is writing the content of the actual lesson.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://carpentries.github.io/sandpaper/}
  \item \url{https://github.com/carpentries/sandpaper/}
  \item \url{https://carpentries.github.io/workbench/}
  \item Report bugs at \url{https://github.com/carpentries/sandpaper/issues/}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@carpentries.org} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

Other contributors:
\itemize{
  \item Julien Colomb (\href{https://orcid.org/0000-0002-3127-5520}{ORCID}) [contributor]
  \item Ben Companjen (\href{https://orcid.org/0000-0002-7023-9047}{ORCID}) [contributor]
  \item Toby Hodges \email{tobyhodges@carpentries.org} [contributor]
  \item François Michonneau \email{francois@carpentries.org} [contributor]
  \item Kelly Barnes \email{kbarnes@carpentries.org} [contributor]
}

}
\keyword{internal}
