% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve}
\alias{serve}
\title{Build your lesson and work on it at the same time}
\usage{
serve(path = ".")
}
\arguments{
\item{path}{the path to your lesson. Defaults to the current path.}
}
\value{
the output of \code{servr::httw()}, invisibly. This is mainly used for its
side-effect
}
\description{
This function will serve your lesson and it will auto-update whenever you
save a file. s
}
\examples{
if (FALSE) {
 # create an example lesson 
 tmp <- tempfile()
 create_lesson(tmp, open = FALSE)
 
 # open the episode for editing
 file.edit(fs::path(tmp, "episodes", "01-introduction.Rmd"))

 # serve the lesson and begin editing the file. Watch how the file will
 # auto-update whenever you save it. 
 sandpaper::serve()
}
}
