% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dropdown.R
\name{set_config}
\alias{set_config}
\title{Set individual keys in a configuration file}
\usage{
set_config(pairs = NULL, create = FALSE, path = ".", write = FALSE)
}
\arguments{
\item{pairs}{a named character vector with keys as the names and the new
values as the contents}

\item{create}{if \code{TRUE}, any new values in \code{pairs} will be created and
appended; defaults to \code{FALSE}, which prevents typos from sneaking in.
single key-pair values currently supported.}

\item{path}{path to the lesson. Defaults to the current directory.}

\item{write}{if \code{TRUE}, the schedule will overwrite the schedule in the
current file.}
}
\description{
Set individual keys in a configuration file
}
\details{
This function deals strictly with keypairs in the yaml. For lists, see
\code{\link[=set_dropdown]{set_dropdown()}}.
\subsection{Default Keypairs Known by Sandpaper}{

When you create a new lesson in sandpaper, the following keys are known:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{carpentry: 'incubator' # [character] one of cp, dc, swc, lab, incubator
title: 'Lesson Title' # [character] the lesson title
created: '2021-02-09' # [character] Date in ISO 8601 format; defaults to ~
keywords: 'static site, lesson template, markdown, R, software' # comma-separated list
life_cycle: 'pre-alpha' # [character]: pre-alpha, alpha, beta, stable
license: 'CC-BY 4.0' # [character]
source: 'https://github.com/carpentries/sandpaper-docs/' # github repository
branch: 'main' # [character] default branch
contact: 'zkamvar@carpentries.org'
}\if{html}{\out{</div>}}
}

\subsection{Optional Keypairs Known by Sandpaper}{

The following keypairs are known by sandpaper, but are optional:

\if{html}{\out{<div class="sourceCode yaml">}}\preformatted{url: 'https://lessons.carpentries.org/sandpaper-docs' # custom URL if you are
  # deploying to a custom url
fail_on_error: yes # [boolean] for R Markdown lessons; fail the build if any
  # chunks produce an error. Use `error: true` in chunk options to allow the 
  # error to be displayed
workbench-beta: yes # [boolean] if truthy, this displays a banner on the site
  # that indicates the site is in the workbench beta phase.
}\if{html}{\out{</div>}}

As the workbench becomes more developed, some of these optional keys may
disappear.
}
}
