% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sidebar.R
\name{create_resources_dropdown}
\alias{create_resources_dropdown}
\alias{create_sidebar}
\alias{update_sidebar}
\title{Create the sidebar for varnish}
\usage{
create_resources_dropdown(files, type = "learners")

create_sidebar(
  chapters,
  name = "",
  html = "<a href='https://carpentries.org'/>"
)

update_sidebar(
  sidebar = NULL,
  nodes = NULL,
  this_page = NULL,
  title = NULL,
  item = NULL
)
}
\arguments{
\item{files}{a vector of markdown file names}

\item{type}{one of "learners" (default) or "instructors". If it is learners,
the setup page will be excluded since it is included in the index. For
"instructors", the instructor notes are included and the learner profiles
are included.}

\item{chapters}{a character vector of paths to markdown chapters}

\item{name}{the name of the current chapter}

\item{html}{the html of the current chapter. defaults to a link that will
produce a sidebar with no links to headings.}

\item{sidebar}{an object of class "list-store" which has a \code{"sidebar"}
element in the stored list. See \code{\link[=set_globals]{set_globals()}}.}

\item{nodes}{the HTML nodes of an HTML page}

\item{this_page}{the path to the current HTML page}

\item{title}{the current title}

\item{item}{the index of the sidebar item to update}
}
\value{
a character vector of HTML divs that can be appended to display the
sidebar.
}
\description{
Varnish uses a sidebar for navigation across and within an episode. Each
episode's sidebar is different because there needs to be a clear indicator
which episode is the current one within the sidebar.
}
\details{
This function creates that sidebar.
}
\seealso{
\code{\link[=sidebar_item]{sidebar_item()}} for creation of individual sidebar items,
\code{\link[=set_globals]{set_globals()}} for where \code{create_sidebar()} is called and
\code{\link[=build_html]{build_html()}} for where \code{update_sidebar()} is called.
}
\keyword{internal}
