% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template_gitignore}
\alias{template_gitignore}
\alias{template_episode}
\alias{template_links}
\alias{template_config}
\alias{template_pkgdown}
\title{Template files}
\usage{
template_gitignore()

template_episode()

template_links()

template_config()

template_pkgdown()
}
\value{
a character string with the path to the template within the
{sandpaper} repo.
}
\description{
Use these files as templates for your own sandpaper lesson
}
\examples{

cat(readLines(template_gitignore(), n = 6), sep = "\n")
}
