% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template_gitignore}
\alias{template_gitignore}
\alias{template_episode}
\alias{template_links}
\alias{template_citation}
\alias{template_config}
\alias{template_conduct}
\alias{template_index}
\alias{template_license}
\alias{template_contributing}
\alias{template_setup}
\alias{template_pkgdown}
\alias{template_placeholder}
\alias{template_pr_diff}
\alias{template_sidebar_item}
\alias{template_metadata}
\title{Template files}
\usage{
template_gitignore()

template_episode()

template_links()

template_citation()

template_config()

template_conduct()

template_index()

template_license()

template_contributing()

template_setup()

template_pkgdown()

template_placeholder()

template_pr_diff()

template_sidebar_item()

template_metadata()
}
\value{
a character string with the path to the template within the
\code{{sandpaper}} repo.
}
\description{
Use these files as templates for your own sandpaper lesson
}
\examples{

cat(readLines(template_gitignore(), n = 6), sep = "\n")
}
\keyword{internal}
