% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_workflows.R
\name{update_github_workflows}
\alias{update_github_workflows}
\title{Update github workflows}
\usage{
update_github_workflows(
  path = ".",
  files = "",
  overwrite = TRUE,
  clean = "*.yaml",
  quiet = FALSE
)
}
\arguments{
\item{path}{path to the current lesson.}

\item{files}{the files to include in the update. Defaults to an empty string,
which will update all files}

\item{overwrite}{if \code{TRUE} (default), the file(s) will be overwritten.}

\item{clean}{glob of files to be cleaned before writing. Defaults to
\code{"*.yaml"}.  to remove all files with the four-letter "yaml" extension
(but it will not remove the ".yml" extension). You can also specify a
whole file name like "workflow.yaml" to remove one specific file. If you
do not want to clean, set this to \code{NULL}.}

\item{quiet}{if \code{TRUE}, the process will not output any messages, default is
\code{FALSE}, which will report on the progress of each step.}
}
\value{
the paths to the new files.
}
\description{
This function copies and updates the workflows to run \code{{sandpaper}}.
}
